requirements.txt
torch
numpy
pillow
torchvision
lightning
facenet_pytorch
opencv-python
kafka-python 
sqlite3
json
pydantic
jose
scipy